chore: remove useless file.
