--- api/test/e2enew/migrate/allData.go
@@ -1,12 +1,0 @@
-package migrate
-
-type AllData struct {
-	Counsumers    []interface{}
-	Routes        []interface{}
-	Services      []interface{}
-	SSLs          []interface{}
-	Upstreams     []interface{}
-	Scripts       []interface{}
-	GlobalPlugins []interface{}
-	PluginConfigs []interface{}
-}
